Moved expand_action
